feat(auth): implement JWT registration, login, controllers and fix security filter chain

+++ src/main/java/com/autodrive/backend/model/Reservation.java
@@ -22,7 +22,7 @@ public class Reservation {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long reservation_id;
+    private Integer reservation_id;
     
     @Column(nullable = false)
     private LocalDate startDate;
